添加Gui跨版本兼容层，`screenAfterInitEvent`方法移动到`ModEvents`，修改coremod检测方式
