+++ common/src/main/java/top/vmctcn/vmtranslationupdate/ModEvents.java
@@ -1,16 +1,22 @@
 package top.vmctcn.vmtranslationupdate;
 
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.client.gui.screen.TitleScreen;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.ClickEvent;
 import net.minecraft.text.HoverEvent;
 import net.minecraft.text.Style;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
+import top.vmctcn.vmtranslationupdate.screen.SuggestModScreen;
 import top.vmctcn.vmtranslationupdate.util.ModConfigUtil;
 
 import top.vmctcn.vmtranslationupdate.util.VersionCheckUtil;
 
 public class ModEvents {
+    public static boolean firstTitleScreenShown = false;
+
     public static void playerJoinEvent(ServerPlayerEntity player) {
         String localVersion = ModConfigUtil.getConfig().modPackTranslationVersion;
         String onlineVersion = VersionCheckUtil.getOnlineVersion();
@@ -35,5 +41,19 @@ public class ModEvents {
                 player.sendMessage(message);
             }
         }
+    }
+
+    public static void screenAfterInitEvent(Screen screen) {
+        if (firstTitleScreenShown || !(screen instanceof TitleScreen)) {
+            return;
+        }
+
+        String language = MinecraftClient.getInstance().getLanguageManager().getLanguage();
+
+        if ("zh_cn".equals(language)) {
+            MinecraftClient.getInstance().setScreen(new SuggestModScreen(screen));
+        }
+
+        firstTitleScreenShown = true;
     }
 }
