+++ common/src/main/java/top/vmctcn/vmtranslationupdate/VMTranslationUpdate.java
@@ -5,7 +5,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import top.vmctcn.vmtranslationupdate.util.GameOptionsUtil;
 import top.vmctcn.vmtranslationupdate.util.ModConfigUtil;
-import top.vmctcn.vmtranslationupdate.util.ScreenUtil;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -17,8 +16,6 @@ public class VMTranslationUpdate {
     public static final Path OPTIONS_FILE = Paths.get(MinecraftClient.getInstance().runDirectory.toString(), "options.txt");
 
     public static void init() {
-        ScreenUtil.checkModsLoaded();
-
         if (ModConfigUtil.getConfig().autoSwitchLanguage && ModConfigUtil.getConfig().switchLanguage != null) {
             try {
                 GameOptionsUtil.createInitFile(OPTIONS_FILE.toFile());
