+++ common/src/main/java/top/vmctcn/vmtranslationupdate/verlayer/ScreenHelper.java
@@ -1,0 +1,16 @@
+package top.vmctcn.vmtranslationupdate.verlayer;
+
+import com.mojang.blaze3d.systems.RenderSystem;
+import net.minecraft.client.gui.DrawContext;
+import net.minecraft.client.render.RenderLayer;
+import net.minecraft.util.Identifier;
+
+public class ScreenHelper {
+    public static void drawGuiTexture(DrawContext context, Identifier sprite, int x, int y, float u, float v, int width, int height, int textureWidth, int textureHeight) {
+        context.drawTexture(RenderLayer::getGuiTextured, sprite, x, y, u, v, width, height, textureWidth, textureHeight);
+    }
+
+    public static void resetShaderColor(DrawContext context) {
+        RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
+    }
+}
