+++ fabric/src/main/java/top/vmctcn/vmtranslationupdate/fabric/VMTranslationUpdateClientFabric.java
@@ -2,8 +2,8 @@ package top.vmctcn.vmtranslationupdate.fabric;
 
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
+import top.vmctcn.vmtranslationupdate.ModEvents;
 import top.vmctcn.vmtranslationupdate.VMTranslationUpdate;
-import top.vmctcn.vmtranslationupdate.util.ScreenUtil;
 
 public class VMTranslationUpdateClientFabric implements ClientModInitializer {
     @Override
@@ -11,7 +11,7 @@ public class VMTranslationUpdateClientFabric implements ClientModInitializer {
         VMTranslationUpdate.init();
 
         ScreenEvents.AFTER_INIT.register((client, screen, scaledWidth, scaledHeight) -> {
-            ScreenUtil.screenAfterInitEvent(screen);
+            ModEvents.screenAfterInitEvent(screen);
         });
     }
 }
